bugfix: Image format error
The server node of the styx package, which is responsible (among other
things) for publishing simulator images to ROS topic /camera/image_raw,
failed to properly set the image's format string. This resulted in
errors when downstream nodes tried to subscribe to the topic, e.g.
rqt_image_view would fail to display images and repeatedly print the
below error message to the console:

ImageView.callback_image() could not convert image from '8UC3' to
'rgb8' ([8UC3] is not a color format. but [rgb8] is. The conversion
does not make sense)

Node source was changed to properly set image format strings.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>