bugfix: Linux Python interpreter can't handle Carriage Return character
The Linux version of the Python interpreter fails to recognize the
Carriage Return ("\r") character, which is used in conjunction with the
New Line ("\n") character to indicate a line break in Windows (in Linux
line breaks are indicated by a single New Line character). This was
causing the interpreter to terminate when loading file
waypoint_updater.py. Fixed by removing all instances of the Carriage
Return character from the file.

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>