feature: Dockerfile for customized ROS image
One of the proposed ways to setup the project environment is to pull a
Docker image from the OSRF repository containing a pre-configured ROS
toolchain. This avoids the need to install basic packages, but manually
adding the extra packages required by the project is still cumbersome.
This commit adds a Dockerfile (and an associated build.sh script for
simple execution) to automated this customization part.

Resolves: #10

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>