bugfix: Error messages from waypoint_updator
The subscriber for traffic light messages expects messages of type
`styx_msgs/TrafficLightArray`, but the actual type of traffic light
messages is `std_msgs/Int32`

Change the type of traffic light message and fix the related issue.

In addition, the `set_waypoint_velocity' function is not properly called
in function `pose_cb'.   

Resolves: #15