feature: Traffic Light Detector reads camera parameters from topic
In order to enable seamless portability between simulated and
real-world environments, the Traffic Light Detector was modified to
read camera parameters from the "/camera_info" topic. Originally this
topic was available only in real-world mode, so simulator settings were
also changed to make this topic available under it as well.

Resolves: #26

Signed-off-by: Helio Perroni Filho <mail@xperroni.me>